Retire Tile to PXA conversion
