+++ pmlc/CMakeLists.txt
@@ -32,7 +32,6 @@ pml_cc_library(
     pmlc::conversion::pxa_to_affine
     pmlc::conversion::stdx_to_llvm
     pmlc::conversion::tile_to_linalg
-    pmlc::conversion::tile_to_pxa
     pmlc::dialect::affinex::transforms
     pmlc::dialect::layer::ir
     pmlc::dialect::layer::transforms
