+++ pmlc/conversion/CMakeLists.txt
@@ -2,4 +2,3 @@ add_subdirectory(linalg_to_pxa)
 add_subdirectory(pxa_to_affine)
 add_subdirectory(stdx_to_llvm)
 add_subdirectory(tile_to_linalg)
-add_subdirectory(tile_to_pxa)
