+++ pmlc/conversion/linalg_to_pxa/CMakeLists.txt
@@ -4,12 +4,13 @@ pml_cc_library(
     passes.h
   SRCS
     linalg_to_pxa.cc
+    tile_to_pxa_special_patterns.cc
+    special_utils.cc
     #pooling.cc
     reshape.cc
     utils.cc
   DEPS
     ::passes-gen
-    pmlc::conversion::tile_to_pxa
     pmlc::dialect::pxa::analysis
     pmlc::dialect::pxa::ir
     pmlc::dialect::stdx::ir
