--- pmlc/conversion/tile_to_pxa/CMakeLists.txt
@@ -1,24 +1,0 @@
-pml_cc_library(
-  NAME tile_to_pxa
-  HDRS
-    passes.h
-  SRCS
-    specials.cc
-    tile_to_pxa.cc
-    utils.cc
-  DEPS
-    ::passes-gen
-    pmlc::dialect::layer::ir
-    pmlc::dialect::pxa::analysis
-    pmlc::dialect::pxa::ir
-    pmlc::dialect::stdx::ir
-    pmlc::dialect::tile::ir
-    pmlc::util
-)
-
-pml_tblgen_library(
-  NAME passes-gen
-  TD_FILE passes.td
-  OUTS
-    -gen-pass-decls passes.h.inc
-)
