--- pmlc/conversion/tile_to_pxa/tests/const.mlir
@@ -1,13 +1,0 @@
-// RUN: pmlc-opt -convert-tile-to-pxa -canonicalize -cse -split-input-file %s | FileCheck %s
-
-func @const() -> tensor<f32> {
-  %cst = tile.constant(3.0 : f64) : tensor<f32>
-  return %cst : tensor<f32>
-}
-
-// CHECK-LABEL: func @const
-// CHECK: constant 3.0{{.*}} : f32
-// CHECK: affine.parallel
-// CHECK:   pxa.reduce assign
-// CHECK:   affine.yield
-// CHECK: return %{{.*}} : memref<f32>
