--- pmlc/conversion/tile_to_pxa/tests/eltwise.mlir
@@ -1,143 +1,0 @@
-// RUN: pmlc-opt -convert-tile-to-pxa -canonicalize -cse %s | FileCheck %s
-
-// CHECK-LABEL: func @bit_not
-func @bit_not(%arg0: tensor<10x20xsi32>) -> tensor<10x20xsi32> {
-  // CHECK: %[[negOne:.*]] = arith.constant -1 : i32
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: subi %[[negOne]], %{{.*}}
-  // CHECK: pxa.reduce assign
-  %0 = tile.bit_not %arg0 : (tensor<10x20xsi32>) -> tensor<10x20xsi32>
-  return %0 : tensor<10x20xsi32>
-}
-
-// CHECK-LABEL: func @neg_i32
-func @neg_i32(%arg0: tensor<10x20xsi32>) -> tensor<10x20xsi32> {
-  // CHECK: %[[c0:.*]] = arith.constant 0 : i32
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: subi %[[c0]], %{{.*}}
-  // CHECK: pxa.reduce assign
-  %0 = tile.neg %arg0 : (tensor<10x20xsi32>) -> tensor<10x20xsi32>
-  return %0 : tensor<10x20xsi32>
-}
-
-// CHECK-LABEL: func @acos_f32
-func @acos_f32(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.acos({{.*}}) : (f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.acos %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @asin_f32
-func @asin_f32(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.asin({{.*}}) : (f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.asin %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @atan_f32
-func @atan_f32(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.atan({{.*}}) : (f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.atan %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @cosh_f32
-func @cosh_f32(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.cosh({{.*}}) : (f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.cosh %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @erf_f32
-func @erf_f32(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.erf({{.*}}) : (f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.erf %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @floor_f32
-func @floor_f32(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.floor({{.*}}) : (f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.floor %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @pow_f32
-func @pow_f32(%arg0: tensor<8x9xf32>, %arg1: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.pow({{.*}}, {{.*}}) : (f32, f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.pow %arg0, %arg1 : (tensor<8x9xf32>, tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @pow_f32_to_si32
-func @pow_f32_to_si32(%arg0: tensor<8x9xf32>, %arg1: tensor<8x9xsi32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.pow({{.*}}, {{.*}}) : (f32, f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.pow %arg0, %arg1 : (tensor<8x9xf32>, tensor<8x9xsi32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @round_f32
-func @round_f32(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.round({{.*}}) : (f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.round %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @sinh_f32
-func @sinh_f32(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.sinh({{.*}}) : (f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.sinh %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @tan_f32
-func @tan_f32(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: stdx.tan({{.*}}) : (f32) -> f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.tan %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
-
-// CHECK-LABEL: func @sin
-func @sin(%arg0: tensor<8x9xf32>) -> tensor<8x9xf32> {
-  // CHECK: affine.parallel
-  // CHECK: pxa.load
-  // CHECK: sin{{.*}} : f32
-  // CHECK: pxa.reduce assign
-  %0 = tile.sin %arg0 : (tensor<8x9xf32>) -> tensor<8x9xf32>
-  return %0 : tensor<8x9xf32>
-}
