+++ pmlc/dialect/pxa/tests/grn.mlir
@@ -1,6 +1,7 @@
 // RUN: pmlc-opt \
 // RUN:   -tile-compute-bounds \
-// RUN:   -convert-tile-to-pxa \
+// RUN:   -convert-tile-to-linalg \
+// RUN:   -convert-linalg-to-pxa \
 // RUN:   -pxa-normalize \
 // RUN:   -canonicalize \
 // RUN:   -pxa-fusion="avoid-reduction-indexes=false" \
